% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchShinyApp.R
\name{launchShinyApp}
\alias{launchShinyApp}
\title{Launch Shiny Application}
\usage{
launchShinyApp()
}
\value{
Runs the shiny application
}
\description{
Connects to Eunomia database. Plots occurrences of selected conditions of interest.
}
